requirements.txt
requests>=2.28.0
pyyaml>=6.0
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
